Add `runOnUI` method to `colorKit`
